DEV: extract DeprecatedIconHandler module
